+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Additional metrics for non-dummy transactions
 - Validation that transaction chain doesn't have any gaps
+### Changed
+- Turn off dummy transactions
 
 ## [v2.18.0] 2020-11-12
 ### Added